folders and files reorganization for popular people page
