--- src/common/Tiles/PersonTile/Title/index.js
@@ -1,7 +1,0 @@
-import { StyledTitle } from "./styled";
-
-const Title = ({ title }) => {
-  return <StyledTitle>{title}</StyledTitle>;
-};
-
-export default Title;
